// @types/FiiListFormattedDataType.d.ts
export type FiiListFormattedDataType = {
  ticker: string;
  category: string;
  price: string;
  dy: string;
  pvp: string;
  caixa: string;
  cagrDividendos3Anos: string;
  cagrDividendos3AnosColor: string;
  cagrValorCota3Anos: string;
  cagrValorCota3AnosColor: string;
  patrimonio: string;
  gestor: string;
  isTopManager: string;
  gestao: string;
  cotistas: string;
  liquidezDiaria: string;
};